fix(description): ensure that props respond
