+++ src/components/Description/index.ts
@@ -1,7 +1,7 @@
 // import { createAsyncComponent } from '/@/utils/factory/createAsyncComponent';
 // export const Description = createAsyncComponent(() => import('./src/index'));
 
-import Description from './src/index';
+import Description from './src/index.vue';
 
 export { Description };
 export * from './src/types';
